Update REE_Chain.bt
Added colors to sections of the chain file. IMPORTANT: I commented out the other bones in the hash list since I'm only working in Rise. You may want to uncomment them or use this as a separate file from this main REE_Chain.bt. I'll leave it up to you. This one may also be missing newer code since it's a tad older.